[flycheck.el] Use shellcheck for sh/bash linting.
Use the popular shellcheck utility for linting and error-checking bash,
but also POSIX sh and zsh as desired.  See the project repo for more
info.

http://www.shellcheck.net/

NOTE: This is used in Arch Linux with the AUR package shellcheck-bin
that pulls the compiled binary for Ubuntu as the Haskell compilation
process possible shellcheck-git is arduous.  See below for more info.

https://aur.archlinux.org/packages/shellcheck-bin